update
						FOOD FOR THE SOUL

WHO   :	Customers who love to listen to music.

WHAT  :	In a digital era where the climate change is at a rapid pace, it’s hard for music lovers to capture lyrics due to frequent music releases in the market where it’s out with the old, in with the new before embracing the last. It brought less focus on lyrics including the now.

WHERE/ WHEN   :	This problem is occurring in a global stand point in modern times.

WHY   :	It’s important to address this issue at hand to enable the ability of embracing food for the soul.